Split module deeplink to api and impl.
