+++ app/build.gradle.kts
@@ -301,6 +301,7 @@ dependencies {
     testImplementation(libs.test.turbine)
     testImplementation(projects.libraries.matrix.test)
     testImplementation(projects.services.toolbox.test)
+    testImplementation(projects.tests.testutils)
 
     koverDependencies()
 }
