+++ libraries/deeplink/api/src/main/kotlin/io/element/android/libraries/deeplink/api/DeeplinkParser.kt
@@ -1,0 +1,14 @@
+/*
+ * Copyright 2023, 2024 New Vector Ltd.
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial
+ * Please see LICENSE files in the repository root for full details.
+ */
+
+package io.element.android.libraries.deeplink.api
+
+import android.content.Intent
+
+fun interface DeeplinkParser {
+    fun getFromIntent(intent: Intent): DeeplinkData?
+}
