--- libraries/deeplink/build.gradle.kts
@@ -1,35 +1,0 @@
-import extension.setupAnvil
-
-/*
- * Copyright 2022-2024 New Vector Ltd.
- *
- * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial
- * Please see LICENSE files in the repository root for full details.
- */
-plugins {
-    id("io.element.android-library")
-}
-
-android {
-    namespace = "io.element.android.libraries.deeplink"
-}
-
-setupAnvil()
-
-dependencies {
-    implementation(projects.libraries.di)
-    implementation(libs.dagger)
-    implementation(libs.androidx.corektx)
-    implementation(projects.libraries.core)
-    implementation(projects.libraries.matrix.api)
-    implementation(projects.libraries.androidutils)
-    implementation(projects.libraries.architecture)
-    implementation(projects.libraries.uiStrings)
-    implementation(projects.services.toolbox.api)
-
-    testImplementation(libs.test.junit)
-    testImplementation(libs.test.truth)
-    testImplementation(libs.test.robolectric)
-    testImplementation(projects.libraries.matrix.test)
-    testImplementation(projects.tests.testutils)
-}
