--- libraries/deeplink/src/main/kotlin/io/element/android/libraries/deeplink/Constants.kt
@@ -1,11 +1,0 @@
-/*
- * Copyright 2023, 2024 New Vector Ltd.
- *
- * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial
- * Please see LICENSE files in the repository root for full details.
- */
-
-package io.element.android.libraries.deeplink
-
-internal const val SCHEME = "elementx"
-internal const val HOST = "open"
